+++ db/schema.ts
@@ -1,0 +1,290 @@
+// public
+import {
+  pgTable,
+  pgSchema,
+  uniqueIndex,
+  serial,
+  varchar,
+  boolean,
+  integer,
+  real,
+  date,
+  time,
+  jsonb,
+} from "drizzle-orm/pg-core";
+import { pointType } from "./PointAdaptation";
+
+export const cities = pgTable(
+  "cities",
+  {
+    id: serial("id").primaryKey(),
+    name: varchar("name").notNull(),
+    state: varchar("state", { length: 2 }),
+  },
+  (cities) => ({
+    nameStateIdx: uniqueIndex("name_state_idx").on(cities.name, cities.state),
+  })
+);
+
+export const researchers = pgTable("researchers", {
+  id: integer("id").primaryKey(),
+  name: varchar("name"),
+});
+
+export const coordinates = pgTable("coordinates", {
+  id: integer("id").primaryKey(),
+  point: pointType("point"),
+  type: varchar("type"),
+});
+
+//CYCLIST COUNT
+
+export const cyclist_count_schema = pgSchema("cyclist_count");
+
+export const cyclist_count_edition = cyclist_count_schema.table("edition", {
+  id: integer("id").primaryKey(),
+  cityId: integer("city_id").references(() => cities.id),
+  name: varchar("name"),
+  date: date("date"),
+  coordinatesId: integer("coordinates_id").references(() => coordinates.id),
+});
+
+export const cyclist_count_summary = cyclist_count_schema.table("summary", {
+  id: integer("id").primaryKey(),
+  editionId: integer("edition_id").references(() => cyclist_count_edition.id),
+  count: integer("count"),
+});
+
+export const cyclist_count_researcherSession = cyclist_count_schema.table(
+  "researcher_session",
+  {
+    countSessionId: integer("count_session_id")
+      .references(() => cyclist_count_session.id)
+      .notNull(),
+    researcherId: integer("researcher_id")
+      .references(() => researchers.id)
+      .notNull(),
+  }
+);
+
+export const cyclist_count_session = cyclist_count_schema.table("session", {
+  id: integer("id").primaryKey(),
+  editionId: integer("edition_id").references(() => cyclist_count_edition.id),
+  startTime: time("start_time"),
+  endTime: time("end_time"),
+});
+
+export const cyclist_count_directionCounts = cyclist_count_schema.table(
+  "direction_counts",
+  {
+    id: integer("id").primaryKey(),
+    sessionId: integer("session_id").references(() => cyclist_count_session.id),
+    origin: varchar("origin"),
+    originCardinal: varchar("origin_cardinal"),
+    destin: varchar("destin"),
+    destinCardinal: varchar("destin_cardinal"),
+    count: integer("count"),
+  }
+);
+
+export const cyclist_count_characteristicsCount = pgTable(
+  "cyclist_count.characteristics_count",
+  {
+    id: integer("id").primaryKey(),
+    sessionId: integer("session_id").references(() => cyclist_count_session.id),
+    characteristicsId: integer("characteristics_id").references(
+      () => cyclist_count_characteristics.id
+    ),
+    count: integer("count"),
+  }
+);
+
+export const cyclist_count_characteristics = cyclist_count_schema.table(
+  "characteristics",
+  {
+    id: integer("id").primaryKey(),
+    name: varchar("name"),
+    type: varchar("type"),
+    atributo: varchar("atributo"),
+  }
+);
+
+// CYCLIST INFRA
+
+export const cyclist_infra_schema = pgSchema("cyclist_infra");
+
+export const cyclist_infra_relations = cyclist_infra_schema.table(
+  "cycling_infra.relations",
+  {
+    osmId: integer("osm_id").primaryKey(),
+    name: varchar("name"),
+    pdcRef: varchar("pdc_ref"),
+    pdcNotes: varchar("pdc_notes"),
+    pdcTypology: varchar("pdc_typology"),
+    kmTotal: real("km_total"),
+    kmTotalCycleway: real("km_total_cycleway"),
+    kmTotalCyclepath: real("km_total_cyclepath"),
+    kmTotalCyclingRoute: real("km_total_cycling_route"),
+    kmTotalFootway: real("km_total_footway"),
+  }
+);
+
+export const cyclist_infra_relationCities = cyclist_infra_schema.table(
+  "cycling_infra.relation_cities",
+  {
+    relationId: integer("relation_id").references(() => cyclist_infra_relations.osmId),
+    citiesId: integer("cities_id").references(() => cities.id),
+  }
+);
+
+export const cyclist_infra_ways = cyclist_infra_schema.table(
+  "cycling_infra.ways",
+  {
+    osmId: integer("osm_id").primaryKey(),
+    name: varchar("name"),
+    length: real("length"),
+    highway: varchar("highway"),
+    hasCycleway: boolean("has_cycleway"),
+    cyclewayTypology: varchar("cycleway_typology"),
+    relationId: integer("relation_id").references(() => cyclist_infra_relations.osmId),
+    geojson: jsonb("geojson"),
+  }
+);
+
+// CYCLIST PROFILE
+
+export const cyclist_profile_schema = pgSchema("cyclist_profile");
+
+export const cyclist_profile_edition = cyclist_profile_schema.table(
+  "edition",
+  {
+    id: serial("id").primaryKey(),
+    year: integer("year").notNull(),
+    cityId: integer("city_id")
+      .references(() => cities.id)
+      .notNull(),
+  },
+  (edition) => ({
+    yearCityIdx: uniqueIndex("year_city_idx").on(edition.year, edition.cityId),
+  })
+);
+
+export const cyclist_profile_categories = cyclist_profile_schema.table(
+  "categories",
+  {
+    id: serial("id").primaryKey(),
+    type: varchar("type"),
+    name: varchar("name"),
+  },
+  (categories) => ({
+    typeNameIdx: uniqueIndex("type_name_idx").on(
+      categories.type,
+      categories.name
+    ),
+  })
+);
+
+export const cyclist_profile_summary = cyclist_profile_schema.table("summary", {
+  id: serial("id").primaryKey(),
+  categoryId: integer("category_id")
+    .references(() => cyclist_profile_categories.id)
+    .notNull(),
+  value: real("value"),
+  total: integer("total"),
+});
+
+export const cyclist_profile_editionSummary = cyclist_profile_schema.table(
+  "edition_summary",
+  {
+    editionId: integer("edition_id")
+      .references(() => cyclist_profile_edition.id)
+      .notNull(),
+    summaryId: integer("summary_id")
+      .references(() => cyclist_profile_summary.id)
+      .notNull(),
+  }
+);
+
+export const cyclist_profile_formData = cyclist_profile_schema.table(
+  "form_data",
+  {
+    id: integer("id").primaryKey(),
+    date: varchar("date"),
+    weekday: varchar("weekday"),
+    coordinatesId: integer("coordinates_id").references(() => coordinates.id),
+    areaId: integer("area_id").references(() => cyclist_profile_categories.id),
+    neighborhood: varchar("neighborhood"),
+    address: varchar("address"),
+    job: varchar("job"),
+    age: integer("age"),
+    ageCategoryId: integer("age_category_id").references(() => cyclist_profile_categories.id),
+    genderId: integer("gender_id").references(() => cyclist_profile_categories.id),
+    schoolingId: integer("schooling_id").references(() => cyclist_profile_categories.id),
+    wageStandardId: integer("wage_standard_id").references(() => cyclist_profile_categories.id),
+    colorRaceId: integer("color_race_id").references(() => cyclist_profile_categories.id),
+    peopleAtHome: integer("people_at_home"),
+    neighborhoodLiving: varchar("neighborhood_living"),
+    weekUsageTotalCatId: integer("week_usage_total_cat_id").references(
+      () => cyclist_profile_categories.id
+    ),
+    weekUsageTotal: integer("week_usage_total"),
+    weekUsageWorking: integer("week_usage_working"),
+    weekUsageSchool: integer("week_usage_school"),
+    weekUsageShopping: integer("week_usage_shopping"),
+    weekUsageLeisure: integer("week_usage_leisure"),
+    weekUsageStation: integer("week_usage_station"),
+    transportationIntegrationId: integer(
+      "transportation_integration_id"
+    ).references(() => cyclist_profile_categories.id),
+    transportationBeforeBicycleId: integer(
+      "transportation_before_bicycle_id"
+    ).references(() => cyclist_profile_categories.id),
+    distanceTimeId: integer("distance_time_id"),
+    distanceTimeCategoryId: integer("distance_time_category_id").references(
+      () => cyclist_profile_categories.id
+    ),
+    yearsUsingId: integer("years_using_id").references(() => cyclist_profile_categories.id),
+    sharedBicycle: boolean("shared_bicycle"),
+    collisionsLastYears: integer("collisions_last_years").references(
+      () => cyclist_profile_categories.id
+    ),
+    neighborhoodOrigin: varchar("neighborhood_origin"),
+    neighborhoodDestiny: varchar("neighborhood_destiny"),
+    shareBikeChanges: boolean("share_bike_changes"),
+    biggestIssue: integer("biggest_issue").references(() => cyclist_profile_categories.id),
+    biggestNeed: integer("biggest_need").references(() => cyclist_profile_categories.id),
+    motivationToStart: integer("motivation_to_start").references(
+      () => cyclist_profile_categories.id
+    ),
+    motivationToContinue: integer("motivation_to_continue").references(
+      () => cyclist_profile_categories.id
+    ),
+    covid19Changes: boolean("covid19_changes"),
+    covid19FrequencyChanges: varchar("covid19_frequency_changes"),
+    covid19AtHome: varchar("covid19_at_home"),
+  }
+);
+
+export const cyclist_profile_editionFormData = pgTable(
+  "cyclist_profile.edition_form_data",
+  {
+    editionId: integer("edition_id")
+      .references(() => cyclist_profile_edition.id)
+      .notNull(),
+    formDataId: integer("form_data_id")
+      .references(() => cyclist_profile_formData.id)
+      .notNull(),
+  }
+);
+
+export const cyclist_profile_researcherFormData = cyclist_profile_schema.table(
+  "researcher_form_data",
+  {
+    formDataId: integer("form_data_id")
+      .references(() => cyclist_profile_formData.id)
+      .notNull(),
+    researcherId: integer("researcher_id")
+      .references(() => researchers.id)
+      .notNull(),
+  }
+);
