--- drizzle/meta/_journal.json
@@ -1,13 +1,0 @@
-{
-  "version": "5",
-  "dialect": "pg",
-  "entries": [
-    {
-      "idx": 0,
-      "version": "5",
-      "when": 1705262506456,
-      "tag": "0000_fast_ogun",
-      "breakpoints": true
-    }
-  ]
-}
