+++ package.json
@@ -6,9 +6,9 @@
   "scripts": {
     "test": "// Sem testes por enquanto",
     "migration:public": "drizzle-kit generate:pg --schema=./db/public_schema.ts",
-    "migration:cyclist_profile": "drizzle-kit generate:pg --schema=./db/cyclist_profile.ts",
-    "migration:cyclist_infra": "drizzle-kit generate:pg --schema=./db/cyclist_infra.ts",
-    "migration:cyclist_count": "drizzle-kit generate:pg --schema=./db/cyclist_count.ts"
+    "migration:cyclist_profile": "drizzle-kit generate:pg --schema=./db/cyclist_profile_schema.ts",
+    "migration:cyclist_infra": "drizzle-kit generate:pg --schema=./db/cyclist_infra_schema.ts",
+    "migration:cyclist_count": "drizzle-kit generate:pg --schema=./db/cyclist_count_schema.ts"
   },
   "repository": {
     "type": "git",
